requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
et-xmlfile==1.1.0
fastapi==0.116.1
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
h11==0.16.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.2
oracledb==2.5.0
packaging==24.2
pandas==2.3.1
pillow==11.1.0
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.1
python-dateutil==2.8.2
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
requests==2.32.4
scipy==1.16.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.1
sympy==1.13.3
thop==0.1.1.post2209072238
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.167
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.35.0
